Fix import in auth handler
The lambda auth handler directly loads the .blueprint.<auth handler>
module directly, bypassing python's normal import mechanisms. This
causes it to skip the __init__.py first. When blueprint then loads
app.py (since its also skipped by the lambda loader), that imports
__init__.py which was importing userpool, which was importing blueprint
causing a circular import only in the auth handler lambda function.

Moving the import inside the function ensures that we lazily do the
import when needed. Which by nessesity means the import is handled after
app.py and the chalice_congito_auth.__init__ are both loaded. This
forces everything to be loaded in the right order and avoid a circular
import.